third_party/WebKit/LayoutTests/media/crash-in-media-moved-to-newdocument.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<iframe width=400 height=300></iframe>
<script>
async_test((t) => {
  var video = document.createElement('video');
  video.src = 'nosuchurl';
  video.play().then(t.unreached_func()
  ).catch(t.step_func_done(function() {
    assert_true(video.paused);
  }));
  window.frames[0].document.body.appendChild(video);
}, "Test for crbug/676004: Crash in blink::beforeCallEnteredCallback()");
</script>
